requirements.txt
crewai[tools]
langgraph
langchain-ollama
ollama
faiss-gpu
pandas
numpy<2.0
PyYAML
requests
duckduckgo-search
beautifulsoup4
tqdm
torch
matplotlib
seaborn
spacy<3.8
textblob
scikit-learn
rich
langsmith
tiktoken
langchain-core
langchain-community
sentence-transformers
